services/recommend-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-generativeai==0.4.1
